refactor(nexblue): centralize command refresh scheduling
