+++ homeassistant/components/nexblue/switch.py
@@ -19,7 +19,6 @@ from .const import DOMAIN
 from .coordinator import NexBlueConfigEntry, NexBlueDataUpdateCoordinator
 
 ASSUMED_STATE_SECONDS = 22
-COMMAND_REFRESH_DELAYS = (3, 20)
 ACTIVE_CHARGING_STATES = frozenset(
     {
         2,  # Charging
@@ -60,9 +59,8 @@ class NexBlueChargingSwitch(
         self._serial_number = serial_number
         self._assumed_is_on: bool | None = None
         self._assumed_state_expires_at = 0.0
-        self._pending_refreshes: set[Callable[[], None]] = set()
         self._assumed_state_expiry_cancel: Callable[[], None] | None = None
-        self.async_on_remove(self._cancel_pending_updates)
+        self.async_on_remove(self._cancel_assumed_state_expiry)
         self._attr_unique_id = f"{serial_number}_charging"
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, serial_number)},
@@ -129,28 +127,12 @@ class NexBlueChargingSwitch(
         self._assumed_is_on = should_charge
         self._assumed_state_expires_at = time.monotonic() + ASSUMED_STATE_SECONDS
         self.async_write_ha_state()
-        self._schedule_command_updates()
+        self._schedule_assumed_state_expiry()
+        self.coordinator.async_schedule_command_refreshes()
 
-    def _schedule_command_updates(self) -> None:
-        """Schedule bounded refreshes and the assumed-state expiry update."""
-        self._cancel_pending_updates()
-
-        def _schedule_refresh(delay: int) -> None:
-            cancel: Callable[[], None] | None = None
-
-            @callback
-            def _request_refresh(_now: datetime) -> None:
-                """Request coordinator data after a command."""
-                if cancel is not None:
-                    self._pending_refreshes.discard(cancel)
-                self.coordinator.config_entry.async_create_task(
-                    self.hass,
-                    self.coordinator.async_request_refresh(),
-                    name="NexBlue command refresh",
-                )
-
-            cancel = async_call_later(self.hass, delay, _request_refresh)
-            self._pending_refreshes.add(cancel)
+    def _schedule_assumed_state_expiry(self) -> None:
+        """Schedule when this switch stops reporting an assumed state."""
+        self._cancel_assumed_state_expiry()
 
         @callback
         def _expire_assumed_state(_now: datetime) -> None:
@@ -159,19 +141,12 @@ class NexBlueChargingSwitch(
             self._assumed_is_on = None
             self.async_write_ha_state()
 
-        for delay in COMMAND_REFRESH_DELAYS:
-            _schedule_refresh(delay)
-
         self._assumed_state_expiry_cancel = async_call_later(
             self.hass, ASSUMED_STATE_SECONDS, _expire_assumed_state
         )
 
-    def _cancel_pending_updates(self) -> None:
-        """Cancel command updates that have not fired."""
-        for cancel in self._pending_refreshes:
-            cancel()
-        self._pending_refreshes.clear()
-
+    def _cancel_assumed_state_expiry(self) -> None:
+        """Cancel the assumed-state expiry callback if it has not fired."""
         if self._assumed_state_expiry_cancel is not None:
             self._assumed_state_expiry_cancel()
             self._assumed_state_expiry_cancel = None
